Extract createOpertaionFetcherFnNodes
